[FIX] Refactoring code in DAG reimplementations
