+++ cuetojson/load.go
@@ -16,18 +16,12 @@ func LoadFile(context *cue.Context, p string, overlay map[string]fs.FS) ([]CuePr
 	}
 
 	for mnt, f := range overlay {
-		f := f
-		mnt := mnt
 		err := fs.WalkDir(f, ".", func(p string, entry fs.DirEntry, err error) error {
 			if err != nil {
 				return err
 			}
 
-			if !entry.Type().IsRegular() {
-				return nil
-			}
-
-			if filepath.Ext(entry.Name()) != ".cue" {
+			if !entry.Type().IsRegular() || filepath.Ext(entry.Name()) != ".cue" {
 				return nil
 			}
 
@@ -45,27 +39,31 @@ func LoadFile(context *cue.Context, p string, overlay map[string]fs.FS) ([]CuePr
 		}
 	}
 
+	return loadCueFiles(context, config, p)
+}
+
+func loadCueFiles(context *cue.Context, config *load.Config, path string) ([]CueProgram, error) {
 	instances := load.Instances([]string{"."}, config)
 	var programs []CueProgram
 
 	if len(instances) == 0 {
-		return nil, fmt.Errorf("Error: Cannot load %s\n", p)
+		return nil, fmt.Errorf("Error: Cannot load %s\n", path)
 	}
 
 	for _, instance := range instances {
 		if instance.Err != nil {
-			fmt.Printf("Error: Cannot load %s\n%s\n", p, instance.Err.Error())
+			fmt.Printf("Error: Cannot load %s\n%s\n", path, instance.Err.Error())
 			return nil, instance.Err
 		}
 		value := context.BuildInstance(instance)
 		if value.Err() != nil {
-			fmt.Printf("Error: Cannot build %s\n%s\n", p, value.Err().Error())
+			fmt.Printf("Error: Cannot build %s\n%s\n", path, value.Err().Error())
 			return nil, value.Err()
 		}
 
 		validation := value.Validate()
 		if validation != nil {
-			fmt.Printf("Error: Cannot validate %s\n%s\n", p, validation.Error())
+			fmt.Printf("Error: Cannot validate %s\n%s\n", path, validation.Error())
 			return nil, value.Err()
 		}
 		programs = append(programs, CueProgram{instance, &value})
